requirements.txt
flask
python-dotenv
requests
onedrivesdk
astropy
astroquery
